extract useZoomOnViewport
